Get rid of facebook_valid_user and facebook_skip_new_user_creation configuration.  I've never understood the valid user option and seems to be somewhat contentious.  The user creation option has been superseeded by ...
Added a facebook_application_user_vivify_chain configuration option that takes an array of method symbols.  Each method is called with the facebook_session in turn until one returns a non-nil, non-false application user.  If none of the vivify methods builds a new application user, then that has the same effect as skipping new user creation in the older configuration.